test(e2e): add basic integration test for GET /ohlc endpoint using Supertest
